backend/app/images.py
from dotenv import load_dotenv
import os 
import cloudinary

load_dotenv()

cloudinary.config(
    cloud_name = os.getenv("CLOUD_NAME"),
    api_key = os.getenv("CLOUDINARY_API_KEY"),
    api_secret = os.getenv("CLOUDINARY_SECRET_KEY"),
    secure = True
)
